+++ src/api/server.module.ts
@@ -17,6 +17,8 @@ import { ChatwootController } from './integrations/chatbot/chatwoot/controllers/
 import { ChatwootService } from './integrations/chatbot/chatwoot/services/chatwoot.service';
 import { DifyController } from './integrations/chatbot/dify/controllers/dify.controller';
 import { DifyService } from './integrations/chatbot/dify/services/dify.service';
+import { FlowiseController } from './integrations/chatbot/flowise/controllers/flowise.controller';
+import { FlowiseService } from './integrations/chatbot/flowise/services/flowise.service';
 import { GenericController } from './integrations/chatbot/generic/controllers/generic.controller';
 import { GenericService } from './integrations/chatbot/generic/services/generic.service';
 import { OpenaiController } from './integrations/chatbot/openai/controllers/openai.controller';
@@ -120,5 +122,8 @@ export const difyController = new DifyController(difyService, prismaRepository, 
 
 const genericService = new GenericService(waMonitor, configService, prismaRepository);
 export const genericController = new GenericController(genericService, prismaRepository, waMonitor);
+
+const flowiseService = new FlowiseService(waMonitor, configService, prismaRepository);
+export const flowiseController = new FlowiseController(flowiseService, prismaRepository, waMonitor);
 
 logger.info('Module - ON');